+++ src/domClass.js
@@ -1,7 +1,0 @@
-import { isFun, isStr, inArray } from './staticMethods.js'
-
-const getClass = (parent, childClass) => {
-    return parent.querySelector(`.${childClass}`)
-}
-
-export default getClass